+++ packages/xchain-binance/src/client.ts
@@ -269,7 +269,7 @@ class Client implements BinanceClient, XChainClient {
    * @param {Asset} asset If not set, it will return all assets available. (optional)
    * @returns {Array<Balance>} The balance of the address.
    */
-  getBalance = async (address?: Address, asset?: Asset): Promise<Balances> => {
+  getBalance = async (address?: Address, assets?: Asset[]): Promise<Balances> => {
     try {
       const balances: BinanceBalances = await this.bncClient.getBalance(address || this.getAddress())
 
@@ -280,7 +280,10 @@ class Client implements BinanceClient, XChainClient {
             amount: assetToBase(assetAmount(balance.free, 8)),
           }
         })
-        .filter((balance) => !asset || assetToString(balance.asset) === assetToString(asset))
+        .filter(
+          (balance) =>
+            !assets || assets.filter((asset) => assetToString(balance.asset) === assetToString(asset)).length,
+        )
     } catch (error) {
       return Promise.reject(error)
     }