+++ packages/xchain-ethereum/CHANGELOG.md
@@ -1,4 +1,9 @@
-# v.x.x.x
+# v.0.12.0 (2021-02-24)
+
+### Breaking change
+
+- Update @xchainjs/xchain-client package to 0.6.0
+- Update `getBalance`
 
 ### Fix