+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,3 +1,11 @@
+# v.0.12.0 (2021-02-24)
+
+### Breaking change
+
+- Update @xchainjs/xchain-client package to 0.6.0
+- Update @xchainjs/xchain-cosmos package to 0.10.0
+- Update `getBalance`
+
 # v.0.11.1 (2021-02-24)
 
 ### Breaking change